Extract out the paths methods and tests
