+++ test/core/linter/linter_test.py
@@ -12,14 +12,12 @@ from sqlfluff.core import FluffConfig, Linter
 from sqlfluff.core.errors import (
     SQLBaseError,
     SQLFluffSkipFile,
-    SQLFluffUserError,
     SQLLexError,
     SQLLintError,
     SQLParseError,
     SQLTemplaterError,
 )
 from sqlfluff.core.linter import LintingResult, runner
-from sqlfluff.core.linter.linter import _find_ignore_config_files
 from sqlfluff.core.linter.runner import get_runner
 from sqlfluff.utils.testing.logging import fluff_log_catcher
 
@@ -39,44 +37,6 @@ def normalise_paths(paths):
     makes them comparable.
     """
     return {pth.replace("/", ".").replace("\\", ".") for pth in paths}
-
-
-def test__linter__path_from_paths__dir():
-    """Test extracting paths from directories."""
-    lntr = Linter()
-    paths = lntr.paths_from_path("test/fixtures/lexer")
-    assert normalise_paths(paths) == {
-        "test.fixtures.lexer.block_comment.sql",
-        "test.fixtures.lexer.inline_comment.sql",
-        "test.fixtures.lexer.basic.sql",
-    }
-
-
-def test__linter__path_from_paths__default():
-    """Test .sql files are found by default."""
-    lntr = Linter()
-    paths = normalise_paths(lntr.paths_from_path("test/fixtures/linter"))
-    assert "test.fixtures.linter.passing.sql" in paths
-    assert "test.fixtures.linter.passing_cap_extension.SQL" in paths
-    assert "test.fixtures.linter.discovery_file.txt" not in paths
-
-
-def test__linter__path_from_paths__exts():
-    """Test configuration of file discovery."""
-    lntr = Linter(
-        config=FluffConfig(overrides={"sql_file_exts": ".txt", "dialect": "ansi"})
-    )
-    paths = normalise_paths(lntr.paths_from_path("test/fixtures/linter"))
-    assert "test.fixtures.linter.passing.sql" not in paths
-    assert "test.fixtures.linter.passing_cap_extension.SQL" not in paths
-    assert "test.fixtures.linter.discovery_file.txt" in paths
-
-
-def test__linter__path_from_paths__file():
-    """Test extracting paths from a file path."""
-    lntr = Linter()
-    paths = lntr.paths_from_path("test/fixtures/linter/indentation_errors.sql")
-    assert normalise_paths(paths) == {"test.fixtures.linter.indentation_errors.sql"}
 
 
 @pytest.mark.parametrize("filesize,raises_skip", [(0, False), (5, True), (2000, False)])
@@ -115,91 +75,6 @@ def test__linter__skip_large_bytes(filesize, raises_skip):
         assert not result
     else:
         assert result
-
-
-def test__linter__path_from_paths__not_exist():
-    """Test that the right errors are raise when a file doesn't exist."""
-    lntr = Linter()
-    with pytest.raises(SQLFluffUserError):
-        lntr.paths_from_path("asflekjfhsakuefhse")
-
-
-def test__linter__path_from_paths__not_exist_ignore():
-    """Test extracting paths from a file path."""
-    lntr = Linter()
-    paths = lntr.paths_from_path("asflekjfhsakuefhse", ignore_non_existent_files=True)
-    assert len(paths) == 0
-
-
-def test__linter__path_from_paths__explicit_ignore():
-    """Test ignoring files that were passed explicitly."""
-    lntr = Linter()
-    paths = lntr.paths_from_path(
-        "test/fixtures/linter/sqlfluffignore/path_a/query_a.sql",
-        ignore_non_existent_files=True,
-        ignore_files=True,
-        working_path="test/fixtures/linter/sqlfluffignore/",
-    )
-    assert len(paths) == 0
-
-
-def test__linter__path_from_paths__sqlfluffignore_current_directory():
-    """Test that .sqlfluffignore in the current directory is read when dir given."""
-    oldcwd = os.getcwd()
-    try:
-        os.chdir("test/fixtures/linter/sqlfluffignore")
-        lntr = Linter()
-        paths = lntr.paths_from_path(
-            "path_a/",
-            ignore_non_existent_files=True,
-            ignore_files=True,
-            working_path="test/fixtures/linter/sqlfluffignore/",
-        )
-        assert len(paths) == 0
-    finally:
-        os.chdir(oldcwd)
-
-
-def test__linter__path_from_paths__dot():
-    """Test extracting paths from a dot."""
-    lntr = Linter()
-    paths = lntr.paths_from_path(".")
-    # Use set theory to check that we get AT LEAST these files
-    assert normalise_paths(paths) >= {
-        "test.fixtures.lexer.block_comment.sql",
-        "test.fixtures.lexer.inline_comment.sql",
-        "test.fixtures.lexer.basic.sql",
-    }
-
-
-def test__config__find_sqlfluffignore_in_same_directory():
-    """Test find ignore file in the same directory as sql file."""
-    ignore_files = _find_ignore_config_files(
-        path="test/fixtures/linter/sqlfluffignore/path_b/query_b.sql",
-        working_path="test/fixtures/linter/sqlfluffignore/",
-    )
-    assert ignore_files == {
-        os.path.abspath("test/fixtures/linter/sqlfluffignore/path_b/.sqlfluffignore"),
-        os.path.abspath("test/fixtures/linter/sqlfluffignore/.sqlfluffignore"),
-    }
-
-
-@pytest.mark.parametrize(
-    "path",
-    [
-        "test/fixtures/linter/sqlfluffignore",
-        "test/fixtures/linter/sqlfluffignore/",
-        "test/fixtures/linter/sqlfluffignore/.",
-    ],
-)
-def test__linter__path_from_paths__ignore(path):
-    """Test extracting paths from a dot."""
-    lntr = Linter()
-    paths = lntr.paths_from_path(path)
-    # We should only get query_b, because of the sqlfluffignore files.
-    assert normalise_paths(paths) == {
-        "test.fixtures.linter.sqlfluffignore.path_b.query_b.sql"
-    }
 
 
 @pytest.mark.parametrize(
@@ -258,10 +133,10 @@ def test__linter__linting_result_check_tuples_by_path(by_path, result_type):
     """Test that a LintingResult can partition violations by the source files."""
     lntr = Linter()
     result = lntr.lint_paths(
-        [
+        (
             "test/fixtures/linter/comma_errors.sql",
             "test/fixtures/linter/whitespace_errors.sql",
-        ]
+        )
     )
     check_tuples = result.check_tuples(by_path=by_path)
     isinstance(check_tuples, result_type)
@@ -308,7 +183,7 @@ def test__linter__linting_result_stats(path, stats):
     https://github.com/sqlfluff/sqlfluff/issues/5673
     """
     lntr = Linter()
-    result = lntr.lint_paths([f"test/fixtures/linter/exit_codes/{path}"])
+    result = lntr.lint_paths((f"test/fixtures/linter/exit_codes/{path}",))
     # NOTE: We're using fake return codes for testing purposes.
     assert result.stats(111, 222) == stats
 
@@ -606,7 +481,7 @@ def test__linter__encoding(fname, config_encoding, lexerror):
             }
         )
     )
-    result = lntr.lint_paths([fname])
+    result = lntr.lint_paths((fname,))
     assert lexerror == (SQLLexError in [type(v) for v in result.get_violations()])
